+++ app/src/test/java/ch/epfl/sdp/database/firebase/api/ServerMockDatabaseAPI.java
@@ -99,7 +99,7 @@ public class ServerMockDatabaseAPI implements ServerDatabaseAPI {
     }
 
     @Override
-    public void sendServerAliveSignal(int signal) {
+    public void sendServerAliveSignal(long signal) {
 
     }
